refactor(studio.giselles.ai): refactor GitHub integration by moving to team integrations and removing old integration page
